AArch64: Port Dalvik to 64bit
Change Dalvik such that it compiles and runs as a 64bit program.
Supports only the C interpreter and builds only on 64bit hosts.
The VM runs ordinary Dalvik bytecode. The interpreter has been
changed to have 64bit registers, but the Java types remain
unchanged. The garbage collector, JNI and interpreter are
functional.

+++ libdex/DexDataMap.cpp
@@ -36,8 +36,8 @@ DexDataMap* dexDataMapAlloc(u4 maxCount) {
     /*
      * Avoiding pulling in safe_iop for safe_iopf.
      */
-    if (!safe_mul(&size, maxCount, sizeof(u4) + sizeof(u2)) ||
-        !safe_add(&size, size, sizeof(DexDataMap))) {
+    if (!safe_mul(&size, size_t(maxCount), sizeof(u4) + sizeof(u2)) ||
+        !safe_add(&size, size_t(size), sizeof(DexDataMap))) {
       return NULL;
     }
